Move prompt logic to its own file
