+++ lib/fuzzy-finder-view.js
@@ -1,10 +1,10 @@
 const {Point, CompositeDisposable} = require('atom')
-const etch = require('etch')
-const $ = etch.dom
+const {renderExperimentPrompt} = require('./experiment-prompt')
 const fs = require('fs-plus')
 const fuzzaldrin = require('fuzzaldrin')
 const fuzzaldrinPlus = require('fuzzaldrin-plus')
 const NativeFuzzy = require('@atom/fuzzy-native')
+const SCORING_SYSTEMS = require('./scoring-systems')
 
 const path = require('path')
 const SelectListView = require('atom-select-list')
@@ -16,16 +16,6 @@ const getIconServices = require('./get-icon-services')
  * Minimum number of files in the project needed to show the fast mode prompt.
  */
 const ThresholdForShowingPrompt = 10000
-
-/**
- * These scoring systems should be synchronized with the enum values
- * on the scoringSystem config option defined in package.json
-**/
-const SCORING_SYSTEMS = {
-  STANDARD: 'standard',
-  ALTERNATE: 'alternate',
-  FAST: 'fast'
-}
 
 const MAX_RESULTS = 10
 
@@ -104,7 +94,7 @@ module.exports = class FuzzyFinderView {
           })
         } else {
           return this.selectListView.update({
-            infoMessage: this.renderExperimentPrompt()
+            infoMessage: renderExperimentPrompt()
           })
         }
       },
@@ -373,54 +363,11 @@ module.exports = class FuzzyFinderView {
     } else {
       return this.selectListView.update({
         items: this.items,
-        infoMessage: this.renderExperimentPrompt(),
+        infoMessage: renderExperimentPrompt(),
         loadingMessage: null,
         loadingBadge: null
       })
     }
-  }
-
-  renderExperimentPrompt () {
-    return (
-      $.span({className: 'experiment-prompt'}, [
-        $.a({
-          onmousedown: this.enableExperimentalFuzzyFinder.bind(this)
-        }, [
-          $.span({className: 'badge badge-info'}, 'NEW'),
-          $.span({className: ''}, 'Try experimental fast mode?'),
-          $.span({className: 'icon icon-microscope'}, '')
-        ])
-      ])
-    )
-  }
-
-  enableExperimentalFuzzyFinder () {
-    const notification = atom.notifications.addInfo(
-      'The project you\'ve opened is quite large, which may cause performance issues on the quick open menu. Do you want to enable the new experimental fast mode for the quick open menu?',
-      {
-        detail: 'This mode can be disabled from the fuzzy finder settings later',
-        dismissable: true,
-        buttons: [
-          {
-            text: 'Enable experimental fast mode',
-            onDidClick: () => {
-              atom.config.set('fuzzy-finder.scoringSystem', SCORING_SYSTEMS.FAST)
-              atom.config.set('fuzzy-finder.useRipGrep', true)
-
-              notification.dismiss()
-            },
-            className: 'btn btn-info btn-primary'
-          },
-          {
-            text: 'No',
-            onDidClick: () => {
-              notification.dismiss()
-            }
-          }
-        ],
-        icon: 'rocket'
-      }
-    )
   }
 
   projectRelativePathsForFilePaths (filePaths) {
