+++ lib/scoring-systems.js
@@ -1,0 +1,9 @@
+/**
+ * These scoring systems should be synchronized with the enum values
+ * on the scoringSystem config option defined in package.json
+**/
+module.exports = {
+  STANDARD: 'standard',
+  ALTERNATE: 'alternate',
+  FAST: 'fast'
+}
